feat: add update batch functionality
This commit introduces batch updates. Currently it is only possible
to update one document at a time, using the id in the URL. With this update,
one can send a request body containing a query and an update, with the update
being applied to all documents that match the supplied query.