fix: allow nested queries using dot notation to pass strings without conversion
allows a query such as { "field._id": "1234567890" } through without converting the
value to an ObjectID if it resembles one (that is, 12 byte string or 24 hex chars)

will only allow this if the first part of the key (e.g. "field") can be loacted
in the schema and discovered to by an "Object" field.

Fix #54